requirements.txt
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.7
django-decouple==2.1
gunicorn==20.0.0
heroku==0.1.4
idna==2.8
Pillow==6.2.1
psycopg2==2.8.4
python-dateutil==1.5
python-decouple==3.3
pytz==2019.3
qrcode==6.1
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==4.1.4
openpyxl==3.0.2
